+++ src/main/java/dev/louis/nebula/api/mana/manager/ManaManager.java
@@ -3,7 +3,7 @@ package dev.louis.nebula.api.mana.manager;
 import dev.louis.nebula.api.mana.storage.ManaStorage;
 
 public interface ManaManager extends ManaStorage {
-    float getMana();
-    float getCapacity();
+    long getMana();
+    long getCapacity();
     void tick();
 }
